requirements.txt
biopython==1.79
cGraph==0.1
click==8.1.3
cycler==0.11.0
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.34.4
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.0
packaging==21.3
Pillow==9.2.0
pomegranate==0.14.8
pyparsing==3.0.9
python-box==6.0.2
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.8.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
